:recycle: refactor: separacao do filtro de vagas e scroll de categorias #131
